Fix API Model

+++ IO.SDK/API/DTO/FuelTankDTO.h
@@ -5,7 +5,7 @@ namespace IO::SDK::API::DTO
     struct FuelTankDTO
     {
         int id;
-        int capacity;
-        int quantity;
+        double capacity;
+        double quantity;
     };
 }
